requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pytesseract==0.3.10
Pillow==10.1.0
streamlit
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.26.2
pandas==2.1.3
easyocr==1.7.1
psutil==5.9.7
email-validator==2.1.0.post1
